support local data
Summary:
Currently we have many users across different team using PySlowFast to train networks on kinetics. At the peak time, we would have more than 2000 GPUs fetching the data from everstore, causing a qps at around 20k, which exceeded the normal capacity supported by everstore on kinetics 400 data fetching.
So we are now moving to support to fetch kinetics raw videos from local disk `/data/local/packages/ai-group.kinetics400/prod/kinetics400` as default setting, which could significantly reduce the network I/O time and speed up our training.

Regarding to the local data, it is located at each of the machine (dev, fblearner, etc) as a pre-installed package, which is the same as what we have for ImageNet. The videos were pre-processed and uploaded as a fb-package (not as catching for everstore or gluster), and the data was using by classy vision.

Reviewed By: feichtenhofer

Differential Revision: D21668337

fbshipit-source-id: 54e4b8f7d484e040c1cf1b249dc6815e66a04c2a