add activation checkpoint to ResNet (#421)
Summary:
- This change uses FairScale's checkpoint_wrapper on ResNet based on profiling data.
- Updated INSTALL.md to mention the dependencies on FairScale.
- A new MODEL.ENABLE_AC is added to control this.
- Updated DDP's argument to support this.
- Tested and it enables batch size == 128 per GPU for 32GB GPUs.

# Profiling Data

We used `LayerwiseMemoryTracker` by QuentinDuval from VISSL project to profile this model.

Initially, without checkpointing, batch size can reach 32 per GPU. Here are three graphs for the memory usage for batch size 8, 16 and 32.

<img width="429" alt="image" src="https://user-images.githubusercontent.com/24926999/119875198-5a7a1e80-bedb-11eb-99a0-271c342bbbcb.png">
<img width="439" alt="image" src="https://user-images.githubusercontent.com/24926999/119875235-636af000-bedb-11eb-8f8d-c585fc96c12c.png">
<img width="459" alt="image" src="https://user-images.githubusercontent.com/24926999/119875268-6b2a9480-bedb-11eb-964c-beef44f1b348.png">

As batch size increase the peak increased to 18GB but the valley (lowest points) stayed the same.

Looking at the graphs on per-layer activation memory, we see that early layers have about 4X more memory usage than the next largest:

<img width="438" alt="image" src="https://user-images.githubusercontent.com/24926999/119875913-2a7f4b00-bedc-11eb-9e00-ddbfc955990d.png">

Therefore, I examined those layers and determined that they are the `s1` and `s2`. So in this change, I combined s1 and s2 in a `nn.Sequential` and checkpointed it when the checkpoint flag is set to True. This way, the activation tensors from them will not be saved in the forward and will be recomputed during the backward pass to reduce the peak.

After this change, the memory usage for bs=8, 64 and 128 per GPU are the following.

<img width="434" alt="image" src="https://user-images.githubusercontent.com/24926999/119875623-d2484900-bedb-11eb-8684-bd4059f91872.png">
<img width="451" alt="image" src="https://user-images.githubusercontent.com/24926999/119875653-da07ed80-bedb-11eb-814e-aaa481ec66a2.png">
<img width="445" alt="image" src="https://user-images.githubusercontent.com/24926999/119875683-e3915580-bedb-11eb-9e85-bcac14c2c1b6.png">

The graphs show that there could be more checkpointed blocks for s3, s4, and s5. I leave those for future tuning. bxiong1202  and lyttonhao, do you think I can show you how to use the tools and you guys can do further tuning on your models since you are much more familiar with the model and code base of this framework?

Pull Request resolved: https://github.com/facebookresearch/SlowFast/pull/421

Reviewed By: bxiong1202

Differential Revision: D29143285

Pulled By: haooooooqi

fbshipit-source-id: b7885db0b65fdbbf3a0730cfefb62a17ba9ef147